Moved hkjs.kt to package arrow
